# api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ValidateErrorJSON": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"UserCreationResponse": {
				"description": "Represents a user object with database timestamps and ID.",
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId",
						"description": "The unique identifier of the user (e.g., MongoDB ObjectId).",
						"example": "60b8f4d5f8a3c3001f3e9a0e"
					},
					"name": {
						"type": "string",
						"description": "The user's name.",
						"example": "John Doe"
					},
					"email": {
						"type": "string",
						"description": "The user's email.",
						"example": "JohnDoe@gmail.com",
						"pattern": "^(.+)@(.+)$"
					},
					"age": {
						"type": "integer",
						"format": "int32",
						"description": "The user's age.",
						"example": 21,
						"minimum": 0
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "The date and time when the user was created.",
						"example": "2024-01-01T12:00:00.000Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "The date and time when the user was last updated.",
						"example": "2024-01-01T13:00:00.000Z"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"age",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreationBody": {
				"description": "Represents a user create body object",
				"properties": {
					"email": {
						"type": "string",
						"description": "The user's email.",
						"example": "JohnDoe@gmail.com",
						"pattern": "^(.+)@(.+)$"
					},
					"name": {
						"type": "string",
						"description": "The user's name.",
						"example": "John Doe"
					},
					"age": {
						"type": "integer",
						"format": "int32",
						"description": "The user's age.",
						"example": 21,
						"minimum": 0
					}
				},
				"required": [
					"email",
					"name",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateBody": {
				"description": "Represents a user update body object",
				"properties": {
					"email": {
						"type": "string",
						"description": "The user's email.",
						"example": "JohnDoe@gmail.com",
						"pattern": "^(.+)@(.+)$"
					},
					"name": {
						"type": "string",
						"description": "The user's name.",
						"example": "John Doe"
					},
					"age": {
						"type": "integer",
						"format": "int32",
						"description": "The user's age.",
						"example": 21,
						"minimum": 0
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "User API",
		"version": "1.0.7",
		"description": "This is a sample server User server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
		"termsOfService": "http://swagger.io/terms/",
		"license": {
			"name": {
				"name": "MIT License",
				"url": "opensource.org"
			}
		},
		"contact": {
			"name": "Joseph Adogeri",
			"email": "johndoe@gmail.com",
			"url": "http://example.com"
		}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreationResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "john doe",
											"email": "johndoe@tsoa.com",
											"age": 21,
											"createdAt": "2023-01-01T10:00:00.000Z",
											"updatedAt": "2023-01-01T11:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Creates a new user in the system.",
				"summary": "Create a new user",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The user details for creation.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationBody",
								"description": "The user details for creation."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": [
											{
												"name": "john doe",
												"email": "johndoe@tsoa.com",
												"age": 21,
												"createdAt": "2023-01-01T10:00:00.000Z",
												"updatedAt": "2023-01-01T11:30:00.000Z"
											},
											{
												"name": "jane doe",
												"email": "janedoe@tsoa.com",
												"age": 22,
												"createdAt": "2023-01-01T10:00:00.000Z",
												"updatedAt": "2023-01-01T11:30:00.000Z"
											}
										]
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Retrieves the a list of existing users.\r\nSupply the unique user ID from either and receive corresponding user details.",
				"summary": "Gets all users.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "The requested user.",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"name": "john doe",
											"email": "johndoe@tsoa.com",
											"age": 21,
											"createdAt": "2023-01-01T10:00:00.000Z",
											"updatedAt": "2023-01-01T11:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user.\r\nSupply the unique user ID and receive corresponding user details.",
				"summary": "Gets a single user.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "60b8f4d5f8a3c3001f3e9a0e"
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "The newly updated user.",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"name": "john doe",
											"email": "johndoe@tsoa.com",
											"age": 21,
											"createdAt": "2023-01-01T10:00:00.000Z",
											"updatedAt": "2023-01-01T11:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Updates a user in the system.",
				"summary": "Updates an existing user",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "60b8f4d5f8a3c3001f3e9a0e"
					}
				],
				"requestBody": {
					"description": "The user details for modification.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateBody",
								"description": "The user details for modification."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "successful deletion.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Removes an existing user in database.\r\nSupply the unique user ID and deletes corresponding user details.",
				"summary": "Deletes a single user.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "60b8f4d5f8a3c3001f3e9a0e"
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development server"
		},
		{
			"url": "http://localhost:3000/api/v1",
			"description": "Test Server"
		},
		{
			"url": "http://prod:8080/api/v1",
			"description": "Production server"
		}
	],
	"tags": [
		{
			"name": "User",
			"description": "Operations related to user management, including creation, retrieval, and updates."
		}
	]
}